Issue where copies of LuxDataFrame contexts were not resulting in a list of brand new VisSpecs (resulted in list of references to original VisSpecs). This resulted into issues in the compiler where a change to one VisSpec also changed the other copies of the VisSpec.  Added copy functions to create a copy of an individual VisSpec and a function to create a copy of a LuxDataFrame's context.
Issue where copies of LuxDataFrame contexts were not resulting in a list of brand new VisSpecs (resulted in list of references to original VisSpecs). This resulted into issues in the compiler where a change to one VisSpec also changed the other copies of the VisSpec.

Added copy functions to create a copy of an individual VisSpec and a function to create a copy of a LuxDataFrame's context.