Update Pandas Executor's handling of colored bar and line charts. (#48)
* Merge remove_column functions, Bugfix in Generalize Action, update test_view

Generalize action was not correctly removing columns or filters from the current view

* Updated Flight SQL Executor Demo and SQL Executor

Fixed issue in SQL Executor where it was not padding filtered bar charts. Updated VisCollection to use the dataframe's executor.
Added script to migrate flights data to Postgresql database for SQL executor demo.

* Addition of VisSpec and Context Copying, Bugfix in Generalize function

Issue where copies of LuxDataFrame contexts were not resulting in a list of brand new VisSpecs (resulted in list of references to  original VisSpecs). This resulted into issues in the compiler where a change to one VisSpec also changed the other copies of the VisSpec.

Added copy functions to create a copy of an individual VisSpec and a function to create a copy of a LuxDataFrame's context.

* Issue where copies of LuxDataFrame contexts were not resulting in a list of brand new VisSpecs (resulted in list of references to original VisSpecs). This resulted into issues in the compiler where a change to one VisSpec also changed the other copies of the VisSpec.  Added copy functions to create a copy of an individual VisSpec and a function to create a copy of a LuxDataFrame's context.

Issue where copies of LuxDataFrame contexts were not resulting in a list of brand new VisSpecs (resulted in list of references to original VisSpecs). This resulted into issues in the compiler where a change to one VisSpec also changed the other copies of the VisSpec.

Added copy functions to create a copy of an individual VisSpec and a function to create a copy of a LuxDataFrame's context.

* Update test_view.py

* Fix Merge issue in LDF

* Fix Merge Issues in Vis and VisSpec

* Update VisSpec copy_spec function

change copy_spec to use __dict__.copy()

added test to test_action to verify context of more than 2 variables works for generalize action

* Update Pandas Executor handling of Colored Bar and Line Charts

Update Pandas Executor method of filling empty rows for colored bar and line charts using Jared's fix.

* Updated test_executor

Former-commit-id: 6148263fd937fa28f3436112432691fcdac52d50