Adjusting Intestingness Scoring for Bar Charts
Tuned up scoring for regular bar charts, discounting factor was pushing scores too close to 0.

Added discounting factor for colored bar charts to lower the interestingness score based on the number of group by and color categories. Charts with a high number of categories are harder to read in general and less useful to users.