Added Better Handling for Lazy Execution
Previously, when a vis had no mark assigned the SQLExecutor would use refresh source to populate the vis' data. However, this causes an issue when the Lux compiler is trying to create colored barcharts in recommendation actions.

Issue is resolved by approaching the problem similarly to the PandasExecutor. Take a sample of the original dataset and set that as the Vis' data. The vis will then have a mark assigned and execution will continue normally.