Updated SQLExecutor to Not Include Null Values in Metadata
Previously if a database table contained rows with null values, the SQLExecutor would include these in the cardinality and unique values lists.

These inclusions would cause issues where the null value was included in barcharts and the interestingness calculations for these charts would break.

Added addition test to SQLExecutor test suite to ensure that the executor was not including null values in its unique_values list. Included with this was an additional script to upload the aug_test dataset which was used for this test.