Created LuxSQLTable Object
Created the LuxSQLTable object to differentiate between Lux' pandas and SQL functionality.

Users can create a LuxSQLTable object and specify which table they would like to connect the object to within the constructor. All of Lux' recommendation functionality is still available, but users will not be able to perform pandas operations on the object as no data is stored locally.