Added _length Parameter to LuxSQLTable
The _length parameter keeps track of the number of datapoints within the table/view that a LuxSQLTable is connected to.

Added a len() function to make this parameter accessible. Unable to override the __len__() function as doing so breaks an assertion within the parent DataFrame class

Updated interestingness.py to use the new len() function