Update LuxSQLTable __len__() and metadata computation
Rather than referencing the _length parameter throughout the code, update and use the LuxSQLTable len() function.

Added _setup_done parameter to the LuxSQLTable. This will check if the initial setup of the table, retrieving and populating attributes, is completed. This will inform which len() function to use, as the parent len() is required while populating the columns of the LuxSQLTable.